Rearrange the implementation and extensively document it.
We dispatch the overflow case from the normal and subnormal cases ahead
of time, which allows to more easily justify the correctness of the
algorithm. In addition, that new dispatch removes the special code path
for `Infinity`.

The special code path for zeros and NaNs is removed. Those values fall
through as the general case. Since the sign `s` of `-0` is then computed
as 1 instead of -1, we compensate by using `s * v` instead of `abs(v)`.
These changes also avoid the need for the external function `isNaN`.

We simplify the subnormal path to use only one multiplication and one
division by a pre-computed constant.

Finally, and perhaps most importantly, we add significant documentation
about the algorithms for the normal and subnormal paths, with---if not
proof---at least an argument and references of why they are correct.

We rearranged the formula of Veltkamp's splitting to match its
formulation in the reference paper, for better understanding, although
the new computation is equivalent to the old one.