Optimize the implementation not to use any external function.
The special code path for zeros and NaNs is removed. They fall through
as the general case.

We avoid `$Number` by using the asm.js idiom `v = +x`.

We avoid `abs` by using `s * v`. This causes `-0` to fall through as
is, but it turns out that it has the desired effect in the end.

We optimize the subnormal path to use only one multiplication and one
division.

We dispatch the overflow case from the normal and subnormal cases ahead
of time, which allows to more easily justify the correctness of the
algorithm.

We hard-code the magical constants to avoid having to look up external
identifiers, ensuring that immediate values are used in the generated
code.

Finally, and perhaps most importantly, we add significant documentation
about the algorithms for the normal and subnormal paths, with---if not
proof---at least an argument and references of why they are correct.